Add an base class for EventMeshTraceService's implementations.;
